refactor(sync): extract conflict resolution logic into dedicated service
Extracted sync conflict resolution logic from PaginatedSyncCommandHandler into a new SyncConflictResolutionService to improve code organization, testability, and maintainability. This change addresses code review feedback regarding separation of concerns and enables better unit testing of conflict resolution scenarios.

The new service handles:
- Deletion conflict resolution with grace period logic
- Recurring task conflict resolution based on planned dates
- Standard timestamp-based conflict resolution
- Copying remote data to existing tasks while preserving IDs

Updated tests to use the new service directly instead of duplicated logic, ensuring accurate coverage of production code.
