+++ src/test/core/application/features/sync/commands/paginated_sync_command_test.dart
@@ -1,13 +1,14 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:whph/core/domain/features/habits/habit_record.dart';
 import 'package:whph/core/domain/features/tasks/task.dart';
-import 'package:acore/acore.dart';
+import 'package:whph/core/application/features/sync/services/sync_conflict_resolution_service.dart';
 
 void main() {
-  group('PaginatedSyncCommand Conflict Resolution Tests', () {
+  group('SyncConflictResolutionService Tests', () {
+    late SyncConflictResolutionService conflictResolutionService;
+
     setUp(() {
-      // Note: In a full implementation, this would set up proper dependencies
-      // For now, we're testing the conflict resolution logic in isolation
+      conflictResolutionService = SyncConflictResolutionService();
     });
 
     group('Habit Record Sync Conflicts', () {
@@ -33,7 +34,7 @@ void main() {
         );
 
         // Act
-        final result = _testConflictResolution(localHabitRecord, remoteHabitRecord);
+        final result = conflictResolutionService.resolveConflict(localHabitRecord, remoteHabitRecord);
 
         // Assert
         expect(result.action, ConflictAction.keepLocal);
@@ -63,7 +64,7 @@ void main() {
         );
 
         // Act
-        final result = _testConflictResolution(localHabitRecord, remoteHabitRecord);
+        final result = conflictResolutionService.resolveConflict(localHabitRecord, remoteHabitRecord);
 
         // Assert
         expect(result.action, ConflictAction.acceptRemote);
@@ -93,7 +94,7 @@ void main() {
         );
 
         // Act
-        final result = _testConflictResolution(localHabitRecord, remoteHabitRecord);
+        final result = conflictResolutionService.resolveConflict(localHabitRecord, remoteHabitRecord);
 
         // Assert
         expect(result.action, ConflictAction.acceptRemote);
@@ -135,7 +136,7 @@ void main() {
         );
 
         // Act
-        final result = _testConflictResolution(localTask, remoteTask);
+        final result = conflictResolutionService.resolveConflict(localTask, remoteTask);
 
         // Assert
         expect(result.action, ConflictAction.keepLocal);
@@ -170,7 +171,7 @@ void main() {
         );
 
         // Act
-        final result = _testConflictResolution(localTask, remoteTask);
+        final result = conflictResolutionService.resolveConflict(localTask, remoteTask);
 
         // Assert
         expect(result.action, ConflictAction.acceptRemote);
@@ -205,7 +206,7 @@ void main() {
         );
 
         // Act
-        final result = _testConflictResolution(localTask, remoteTask);
+        final result = conflictResolutionService.resolveConflict(localTask, remoteTask);
 
         // Assert
         expect(result.action, ConflictAction.acceptRemote);
@@ -238,7 +239,7 @@ void main() {
         );
 
         // Act
-        final result = _testConflictResolution(localTask, remoteTask);
+        final result = conflictResolutionService.resolveConflict(localTask, remoteTask);
 
         // Assert
         // Should fall back to standard timestamp resolution
@@ -268,7 +269,7 @@ void main() {
         );
 
         // Act
-        final result = _testConflictResolution(localHabitRecord, remoteHabitRecord);
+        final result = conflictResolutionService.resolveConflict(localHabitRecord, remoteHabitRecord);
 
         // Assert
         expect(result.action, ConflictAction.keepLocal);
@@ -295,7 +296,7 @@ void main() {
         );
 
         // Act
-        final result = _testConflictResolution(localHabitRecord, remoteHabitRecord);
+        final result = conflictResolutionService.resolveConflict(localHabitRecord, remoteHabitRecord);
 
         // Assert
         expect(result.action, ConflictAction.acceptRemote);
@@ -321,148 +322,13 @@ void main() {
         );
 
         // Act
-        final result = _testConflictResolution(localHabitRecord, remoteHabitRecord);
+        final result = conflictResolutionService.resolveConflict(localHabitRecord, remoteHabitRecord);
 
         // Assert
         expect(result.action, ConflictAction.acceptRemoteForceUpdate);
         expect(result.reason, contains('identical'));
       });
     });
-  });
-}
-
-// SYNC CONFLICT RESOLUTION TESTING ARCHITECTURE
-//
-// This test suite covers different types of sync conflicts:
-//
-// 1. SAME-ID CONFLICTS (tested here):
-//    - When the same entity is modified on different devices before sync
-//    - Handled by _resolveConflict() method during updateSync processing
-//    - Tests use duplicate logic due to complex dependency injection requirements
-//
-// 2. DIFFERENT-ID CONFLICTS (handled in production):
-//    - When devices independently create instances for the same recurring task
-//    - Handled by _checkForRecurringTaskDuplicate() during createSync processing
-//    - Not covered in these tests due to integration complexity
-//
-// KNOWN LIMITATION: This test duplicates production logic due to complex dependency injection requirements
-//
-// FUTURE IMPROVEMENT: To properly test production code, one of these approaches should be used:
-// 1. Extract conflict resolution into a separate service class with minimal dependencies
-// 2. Create comprehensive mocks for all PaginatedSyncCommandHandler dependencies
-// 3. Use a test harness pattern with dependency injection
-//
-// The current implementation tests the logic thoroughly but doesn't guarantee
-// it exactly matches the production behavior if changes are made only to production.
-
-// Helper method that matches the production conflict resolution logic
-// This should be kept in sync with PaginatedSyncCommandHandler.resolveConflictForTesting()
-ConflictResolutionResult<T> _testConflictResolution<T extends BaseEntity<String>>(T localEntity, T remoteEntity) {
-  final DateTime localTimestamp = localEntity.modifiedDate ?? localEntity.createdDate;
-  final DateTime remoteTimestamp = remoteEntity.modifiedDate ?? remoteEntity.createdDate;
-
-  final bool localIsDeleted = localEntity.deletedDate != null;
-  final bool remoteIsDeleted = remoteEntity.deletedDate != null;
-
-  // Handle deletion conflicts specially (matches production implementation)
-  if (localIsDeleted != remoteIsDeleted) {
-    const Duration deletionGracePeriod = Duration(minutes: 5);
-
-    if (localIsDeleted && !remoteIsDeleted) {
-      if (localTimestamp.difference(remoteTimestamp) > deletionGracePeriod) {
-        return ConflictResolutionResult(
-          action: ConflictAction.keepLocal,
-          winningEntity: localEntity,
-          reason: 'Local deletion ($localTimestamp) occurred significantly after remote modification ($remoteTimestamp)',
-        );
-      } else {
-        return ConflictResolutionResult(
-          action: ConflictAction.acceptRemote,
-          winningEntity: remoteEntity,
-          reason: 'Preferring non-deleted remote entity over recent local deletion (deletion time: $localTimestamp, remote time: $remoteTimestamp)',
-        );
-      }
-    } else if (remoteIsDeleted && !localIsDeleted) {
-      if (remoteTimestamp.difference(localTimestamp) > deletionGracePeriod) {
-        return ConflictResolutionResult(
-          action: ConflictAction.acceptRemote,
-          winningEntity: remoteEntity,
-          reason: 'Remote deletion ($remoteTimestamp) occurred significantly after local modification ($localTimestamp)',
-        );
-      } else {
-        return ConflictResolutionResult(
-          action: ConflictAction.keepLocal,
-          winningEntity: localEntity,
-          reason: 'Preferring non-deleted local entity over recent remote deletion (deletion time: $remoteTimestamp, local time: $localTimestamp)',
-        );
-      }
-    }
-  }
-
-  // Handle recurring task duplication conflicts (matches production implementation)
-  if (localEntity is Task && remoteEntity is Task) {
-    if (localEntity.recurrenceParentId != null &&
-        remoteEntity.recurrenceParentId != null &&
-        localEntity.recurrenceParentId == remoteEntity.recurrenceParentId) {
-      final DateTime? localPlannedDate = localEntity.plannedDate;
-      final DateTime? remotePlannedDate = remoteEntity.plannedDate;
-
-      if (localPlannedDate != null && remotePlannedDate != null) {
-        if (localPlannedDate.isBefore(remotePlannedDate)) {
-          return ConflictResolutionResult(
-            action: ConflictAction.keepLocal,
-            winningEntity: localEntity,
-            reason: 'Local recurring task instance has earlier planned date ($localPlannedDate vs $remotePlannedDate)',
-          );
-        } else if (remotePlannedDate.isBefore(localPlannedDate)) {
-          return ConflictResolutionResult(
-            action: ConflictAction.acceptRemote,
-            winningEntity: remoteEntity,
-            reason: 'Remote recurring task instance has earlier planned date ($remotePlannedDate vs $localPlannedDate)',
-          );
-        }
-      }
-    }
-  }
-
-  // Standard timestamp-based conflict resolution (matches production implementation)
-  if (localTimestamp.isAfter(remoteTimestamp)) {
-    return ConflictResolutionResult(
-      action: ConflictAction.keepLocal,
-      winningEntity: localEntity,
-      reason: 'Local timestamp ($localTimestamp) is newer than remote ($remoteTimestamp)',
-    );
-  } else if (remoteTimestamp.isAfter(localTimestamp)) {
-    return ConflictResolutionResult(
-      action: ConflictAction.acceptRemote,
-      winningEntity: remoteEntity,
-      reason: 'Remote timestamp ($remoteTimestamp) is newer than local ($localTimestamp)',
-    );
-  } else {
-    return ConflictResolutionResult(
-      action: ConflictAction.acceptRemoteForceUpdate,
-      winningEntity: remoteEntity,
-      reason: 'Timestamps are identical ($localTimestamp), preferring remote version for consistency',
-    );
-  }
-}
-
-// Test helper types that mirror the production implementation
-enum ConflictAction {
-  keepLocal,
-  acceptRemote,
-  acceptRemoteForceUpdate,
-}
-
-class ConflictResolutionResult<T> {
-  final ConflictAction action;
-  final T winningEntity;
-  final String reason;
-
-  ConflictResolutionResult({
-    required this.action,
-    required this.winningEntity,
-    required this.reason,
   });
 }
